requirements.txt
pandas==1.4.2
transformers==4.18.0
datasets==2.0.0
seqeval==1.2.2
hydra-core==1.1.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
termcolor==1.1.0